+++ lib/msf/ui/console/command_dispatcher/exploit.rb
@@ -86,6 +86,28 @@ class Exploit
   end
 
   #
+  # Tab completion for the run command
+  #
+  def cmd_run_tabs(str, words)
+    if str.end_with?("=")
+      option_name = str.chop()
+      ::Readline.completion_append_character = " "
+      # require "pry"; binding.pry
+      return tab_complete_option_values(option_name, words, opt: option_name).map { |value| "#{str}#{value}"}
+    else
+      if str.include?("=")
+        str_split = str.split("=")
+        option_value = str_split[1].strip
+        option_name = str_split[0].strip
+        ::Readline.completion_append_character = " "
+        return tab_complete_option_values(option_value, words, opt: option_name).map { |value| "#{option_name}=#{value}"}
+      end
+    end
+    ::Readline.completion_append_character = ''
+    return tab_complete_option_names(str, words).map { |name| "#{name}=" }
+  end
+
+  #
   # Launches exploitation attempts.
   #
   def cmd_exploit(*args)
