+++ lib/msf/ui/console/command_dispatcher/payload.rb
@@ -187,24 +187,71 @@ module Msf
             true
           end
 
+
+          # require "pry"; binding.pry
+          #
+          # Tab completion for the run command
+          #
+          # def cmd_run_tabs(str, words)
+          #   require "pry"; binding.pry
+          #   if str.end_with?("=")
+          #     option_name = str.chop()
+          #     ::Readline.completion_append_character = " "
+          #     # require "pry"; binding.pry
+          #     return tab_complete_option_values(option_name, words, opt: option_name).map { |value| "#{str}#{value}"}
+          #   else
+          #     if str.include?("=")
+          #       str_split = str.split("=")
+          #       option_value = str_split[1].strip
+          #       option_name = str_split[0].strip
+          #       ::Readline.completion_append_character = " "
+          #       return tab_complete_option_values(option_value, words, opt: option_name).map { |value| "#{option_name}=#{value}"}
+          #     end
+          #   end
+          #   ::Readline.completion_append_character = ''
+          #   return tab_complete_option_names(str, words).map { |name| "#{name}=" }
+          # end
+
           def cmd_generate_tabs(str, words)
-            fmt = {
-              '-b' => [ true                                              ],
-              '-E' => [ nil                                               ],
-              '-e' => [ framework.encoders.map { |refname, mod| refname } ],
-              '-h' => [ nil                                               ],
-              '-o' => [ true                                              ],
-              '-P' => [ true                                              ],
-              '-S' => [ true                                              ],
-              '-f' => [ :file                                             ],
-              '-t' => [ @@supported_formats                               ],
-              '-p' => [ true                                              ],
-              '-k' => [ nil                                               ],
-              '-x' => [ :file                                             ],
-              '-i' => [ true                                              ],
-              '-v' => [ nil                                               ]
-            }
-            tab_complete_generic(fmt, str, words)
+            # require "pry"; binding.pry
+
+            if str.start_with?("-")
+              fmt = {
+                  '-b' => [ true                                              ],
+                  '-E' => [ nil                                               ],
+                  '-e' => [ framework.encoders.map { |refname, mod| refname } ],
+                  '-h' => [ nil                                               ],
+                  '-o' => [ true                                              ],
+                  '-P' => [ true                                              ],
+                  '-S' => [ true                                              ],
+                  '-f' => [ :file                                             ],
+                  '-t' => [ @@supported_formats                               ],
+                  '-p' => [ true                                              ],
+                  '-k' => [ nil                                               ],
+                  '-x' => [ :file                                             ],
+                  '-i' => [ true                                              ],
+                  '-v' => [ nil                                               ]
+              }
+              tab_complete_generic(fmt, str, words)
+
+            else
+              if str.end_with?("=")
+                option_name = str.chop()
+                ::Readline.completion_append_character = " "
+                # require "pry"; binding.pry
+                return tab_complete_option_values(option_name, words, opt: option_name).map { |value| "#{str}#{value}"}
+              else
+                if str.include?("=")
+                  str_split = str.split("=")
+                  option_value = str_split[1].strip
+                  option_name = str_split[0].strip
+                  ::Readline.completion_append_character = " "
+                  return tab_complete_option_values(option_value, words, opt: option_name).map { |value| "#{option_name}=#{value}"}
+                end
+              end
+            ::Readline.completion_append_character = ''
+            return tab_complete_option_names(str, words).map { |name| "#{name}=" }
+            end
           end
         end
       end
